ac 268

+++ 268.py
@@ -25,7 +25,7 @@ class Solution:
 
     def missingNumber(self, nums):
         n = len(nums)
-        return n * (n+1) / 2 - sum(nums)
+        return int(n * (n+1) / 2 - sum(nums))
